fixed error with area, updated params and updated readme

+++ headers/constants.hpp
@@ -58,13 +58,13 @@ extern double VIRUS_LIFETIME;
 extern double VIRUS_SPEED;
 extern double INNATE_LIFETIME;
 extern double INNATE_SPEED;
-extern int INNATE_RANGE;               // How far away can it kill viruses/cells
+extern double INNATE_RANGE;               // How far away can it kill viruses/cells
 extern double INNATE_KILL_VIRUS_PROB;  // Chance to kill a virus nearby
 extern double INNATE_KILL_CELL_PROB;   // Chance to kill an infected cell nearby
 extern double INNATE_RECRUIT_PROB;     // Chance to recruit another innate cell
 extern double ANTIBODY_LIFETIME;
 extern double ANTIBODY_SPEED;
-extern int ANTIBODY_RANGE;  // The range at which viruses can be destroyed
+extern double ANTIBODY_RANGE;  // The range at which viruses can be destroyed
 extern double ANTIBODY_KILL_PROB;  // Chance to kill when in range
 extern double IFN_SPEED;
